+++ src/LaravelModulesArchServiceProvider.php
@@ -5,13 +5,18 @@ declare(strict_types=1);
 namespace LaravelModulesArch;
 
 use Illuminate\Support\ServiceProvider;
+use LaravelModulesArch\Console\Commands\MakeActionCommand;
+use LaravelModulesArch\Console\Commands\MakeDtoCommand;
 use LaravelModulesArch\Console\Commands\MakeEntityCommand;
 use LaravelModulesArch\Console\Commands\MakeEnumCommand;
 use LaravelModulesArch\Console\Commands\MakeEventCommand;
 use LaravelModulesArch\Console\Commands\MakeExceptionCommand;
 use LaravelModulesArch\Console\Commands\MakeModuleCommand;
 use LaravelModulesArch\Console\Commands\MakeRepositoryCommand;
+use LaravelModulesArch\Console\Commands\MakeStateCommand;
+use LaravelModulesArch\Console\Commands\MakeValidatorCommand;
 use LaravelModulesArch\Console\Commands\MakeValueObjectCommand;
+use LaravelModulesArch\Console\Commands\MakeViewModelCommand;
 
 class LaravelModulesArchServiceProvider extends ServiceProvider
 {
@@ -47,6 +52,11 @@ class LaravelModulesArchServiceProvider extends ServiceProvider
             MakeEventCommand::class,
             MakeExceptionCommand::class,
             MakeRepositoryCommand::class,
+            MakeActionCommand::class,
+            MakeDtoCommand::class,
+            MakeViewModelCommand::class,
+            MakeStateCommand::class,
+            MakeValidatorCommand::class,
         ];
     }
 }